+++ plugins/conference/lib/api/filters/KalturaConferenceEntryServerNodeFilter.php
@@ -1,12 +1,12 @@
 <?php
 /**
- * @package plugins.konference
+ * @package plugins.conference
  * @subpackage api.filters
  */
 class KalturaConferenceEntryServerNodeFilter extends KalturaConferenceEntryServerNodeBaseFilter
 {
 	public function __construct()
 	{
-		$this->serverTypeIn = array(KonferencePlugin::getCoreValue('ConferenceEntryServerNodeType'));
+		$this->serverTypeIn = array(ConferencePlugin::getCoreValue('EntryServerNodeType', ConferenceEntryServerNodeType::CONFERENCE_ENTRY_SERVER ));
 	}
 }