+++ plugins/conference/lib/api/filters/base/KalturaConferenceEntryServerNodeBaseFilter.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * @package plugins.konference
+ * @package plugins.conference
  * @subpackage api.filters.base
  * @abstract
  */